+++ src/withHooks/useAvatar/useAvatar.ts
@@ -1,0 +1,5 @@
+import { options } from '../../withComponents/AvatarSelect/options';
+
+export function useAvatar(avatarName: string) {
+  return options.find(opt => opt.name === avatarName);
+}